refactor(react-doctor): EditFieldDialog consome os compartilhados de schema (#354)
EditFieldDialog sai do no-giant-component (546 → ~290 linhas):

- useEditFieldForm + initialFromField + PendingSuggestion movidos para
  stats/useEditFieldForm.ts (supressão inline de
  rerender-state-only-in-handlers migrada junto); PendingSuggestion
  re-exportado pelo EditFieldDialog para não quebrar consumidores
- bloco de subcampos substituído pelo SubfieldsEditor compartilhado, com
  adaptador patch→setters (null→[], undefined→'all' — equivalentes: o save
  já materializa hasSubfields ? rule : undefined)
- prompt LLM via JustificationPromptField; pendingRemoval via
  useOptionRemovalGuard; TYPE_LABELS importado de field-labels

Com isso todo bloco de config é o MESMO componente nos dois editores — o
invariante do CLAUDE.md (EditFieldDialog espelha o FieldCard) vale por
construção.
